--- Data/Player.cs
@@ -1,9 +1,0 @@
-using Microsoft.AspNetCore.Identity;
-
-namespace Cribbly.Data
-{
-    public class Player : IdentityUser
-    {
-        public Team Team { get; set; }
-    }
-}
